Updating code to apply performance level distributions. Updating code to allow floats in the demographic csv file
Former-commit-id: 99d3f1388c81248becadbd9b8384e100bb586c43

+++ data_gen/DataGeneration/src/demographics.py
@@ -136,7 +136,7 @@ class Demographics(object):
         '''
         ret_list = [row_list[H_GROUPING], row_list[H_TOTAL], row_list[H_PERF_1], row_list[H_PERF_2], row_list[H_PERF_3], row_list[H_PERF_4]]
         # convert to ints and return
-        return [int(x) for x in ret_list]
+        return [float(x) for x in ret_list]
 
 
 class DemographicStatus(object):
